Migrate tool nodes in its own folder
