+++ Cargo.toml
@@ -29,7 +29,8 @@ members = [
     "libraries/shared-memory-server",
     "libraries/extensions/download",
     "libraries/extensions/telemetry/*",
-    "libraries/extensions/dora-record",
+    "tool_nodes/dora-record",
+    "tool_nodes/dora-rerun",
     "libraries/extensions/ros2-bridge",
     "libraries/extensions/ros2-bridge/msg-gen",
     "libraries/extensions/ros2-bridge/python",
