+++ tool_nodes/dora-rerun/Cargo.toml
@@ -13,3 +13,4 @@ dora-node-api = { workspace = true, features = ["tracing"] }
 eyre = "0.6.8"
 tokio = { version = "1.36.0", features = ["rt"] }
 rerun = { version = "0.15.1", features = ["web_viewer", "image"] }
+ndarray = "0.15.6"
